+++ threeML/plugins/EventListLike.py
@@ -9,8 +9,8 @@ from threeML.exceptions.custom_exceptions import deprecated
 from threeML.io.file_utils import file_existing_and_readable
 from threeML.io.plotting.light_curve_plots import binned_light_curve_plot
 from threeML.plugins.OGIPLike import OGIPLike
-from threeML.utils.OGIP import PHAII
-from threeML.utils.OGIP import PHAWrite
+from threeML.utils.OGIP.pha import PHAII
+from threeML.utils.OGIP.pha import PHAWrite
 from threeML.utils.stats_tools import Significance
 
 __instrument_name = "Generic EventList data"
