+++ threeML/plugins/FermiGBMTTELike.py
@@ -9,7 +9,7 @@ import re
 
 from threeML.exceptions.custom_exceptions import deprecated
 from threeML.plugins.EventListLike import EventListLike
-from threeML.utils.OGIP import InstrumentResponseSet
+from threeML.utils.OGIP.response import InstrumentResponseSet
 from threeML.utils.fermi_relative_mission_time import compute_fermi_relative_mission_times
 from threeML.utils.time_series.event_list import EventListWithDeadTime
