+++ threeML/plugins/OGIPLike.py
@@ -4,8 +4,8 @@ from astromodels.utils.valid_variable import is_valid_variable_name
 from threeML.plugins.DispersionSpectrumLike import DispersionSpectrumLike
 from threeML.plugins.SpectrumLike import SpectrumLike
 from threeML.plugins.XYLike import XYLike
-from threeML.utils.OGIP import PHAII
-from threeML.utils.OGIP import PHAWrite
+from threeML.utils.OGIP.pha import PHAII
+from threeML.utils.OGIP.pha import PHAWrite
 from threeML.utils.spectrum.pha_spectrum import PHASpectrum
 
 __instrument_name = "All OGIP-compliant instruments"
