+++ threeML/test/test_spectrum_class.py
@@ -5,7 +5,7 @@ from astromodels import Powerlaw, PointSource, Model
 
 from threeML.plugins.DispersionSpectrumLike import DispersionSpectrumLike
 from threeML.plugins.SpectrumLike import SpectrumLike
-from threeML.utils.OGIP import OGIPResponse
+from threeML.utils.OGIP.response import OGIPResponse
 from threeML.utils.spectrum.binned_spectrum import BinnedSpectrum, BinnedSpectrumWithDispersion, ChannelSet
 
 __this_dir__ = os.path.join(os.path.abspath(os.path.dirname(__file__)))
