+++ threeML/test/test_spectrumlike.py
@@ -6,7 +6,7 @@ from threeML import JointLikelihood, DataList
 from threeML.io.package_data import get_path_of_data_file
 from threeML.plugins.DispersionSpectrumLike import DispersionSpectrumLike
 from threeML.plugins.SpectrumLike import SpectrumLike
-from threeML.utils.OGIP import OGIPResponse
+from threeML.utils.OGIP.response import OGIPResponse
 
 
 def test_assigning_source_name():
