+++ threeML/utils/data_builders/time_series_builder.py
@@ -10,8 +10,8 @@ from threeML.io.progress_bar import progress_bar
 from threeML.plugins.DispersionSpectrumLike import DispersionSpectrumLike
 from threeML.plugins.OGIPLike import OGIPLike
 from threeML.plugins.SpectrumLike import SpectrumLike, NegativeBackground
-from threeML.utils.OGIP import InstrumentResponse, InstrumentResponseSet, OGIPResponse
-from threeML.utils.OGIP import PHAWrite
+from threeML.utils.OGIP.response import InstrumentResponse, InstrumentResponseSet, OGIPResponse
+from threeML.utils.OGIP.pha import PHAWrite
 from threeML.utils.data_builders.fermi.gbm_data import GBMTTEFile, GBMCdata
 from threeML.utils.data_builders.fermi.lat_data import LLEFile
 from threeML.utils.spectrum.binned_spectrum import BinnedSpectrum, BinnedSpectrumWithDispersion
